+++ src/nova/trame/view/components/file_upload.py
@@ -1,12 +1,13 @@
 """View implementation for FileUpload."""
 
-from typing import Any, List, Tuple, Union
+from typing import Any, Iterable, Tuple, Union
 
 from trame.app import get_server
 from trame.widgets import vuetify3 as vuetify
 from trame_server.core import State
 
 from nova.trame._internal.utils import get_state_param
+from nova.trame.utils.types import TrameTuple
 
 from .remote_file_input import RemoteFileInput
 
@@ -17,10 +18,11 @@ class FileUpload(vuetify.VBtn):
     def __init__(
         self,
         v_model: Union[str, Tuple],
-        base_paths: Union[List[str], Tuple, None] = None,
-        extensions: Union[List[str], Tuple, None] = None,
+        base_paths: Union[Iterable[str], str, None] = None,
+        extensions: Union[Iterable[str], str, None] = None,
         label: str = "",
         return_contents: Union[bool, Tuple] = True,
+        show_server_files: Union[bool, Tuple] = True,
         use_bytes: Union[bool, Tuple] = False,
         **kwargs: Any,
     ) -> None:
@@ -35,15 +37,21 @@ class FileUpload(vuetify.VBtn):
             The state variable to set when the user uploads their file. The state variable will contain a latin1-decoded
             version of the file contents. If your file is binary or requires a different string encoding, then you can
             call `encode('latin1')` on the file contents to get the underlying bytes.
-        base_paths: Union[List[str], Tuple], optional
-            Passed to :ref:`RemoteFileInput <api_remotefileinput>`.
-        extensions: Union[List[str], Tuple], optional
-            Restricts the files shown to the user to files that end with one of the strings in the list.
+        base_paths: Union[Iterable[str], str], optional
+            Passed to :ref:`RemoteFileInput <api_remotefileinput>`. Typical Trame binding syntax doesn't work here as
+            tuples are interpreted as literal extensions to filter with. Instead, you can pass a string with a
+            JavaScript expression to bind this parameter.
+        extensions: Union[Iterable[str], str], optional
+            Restricts the files shown to the user to files that end with one of the strings in the list. Typical Trame
+            binding syntax doesn't work here as tuples are interpreted as literal extensions to filter with. Instead,
+            you can pass a string with a JavaScript expression to bind this parameter.
         label : str, optional
             The text to display on the upload button.
         return_contents : Union[bool, Tuple], optional
             If true, the file contents will be stored in v_model. If false, a file path will be stored in v_model.
             Defaults to true.
+        show_server_files : Union[bool, Tuple], optional
+            If true, then the "From Server" option to select a file will be shown. Defaults to true.
         use_bytes : Union[bool, Tuple], optional
             If true, then files uploaded from the local machine will contain bytes rather than text.
         **kwargs
@@ -61,9 +69,12 @@ class FileUpload(vuetify.VBtn):
         self._extensions = extensions if extensions else []
         self._return_contents = return_contents
         self._use_bytes = use_bytes
+        self._show_server_files = TrameTuple.create(show_server_files)
         self._ref_name = f"nova__fileupload_{self._next_id}"
 
-        super().__init__(label, **kwargs)
+        super().__init__(
+            label, click=f"!{self._show_server_files.expression} && trame.refs.{self._ref_name}.click()", **kwargs
+        )
         self.create_ui()
 
     @property
@@ -74,7 +85,11 @@ class FileUpload(vuetify.VBtn):
         self.local_file_input = vuetify.VFileInput(
             v_model=self._v_model,
             __properties=["accept"],
-            accept=",".join(self._extensions) if isinstance(self._extensions, list) else self._extensions,
+            accept=",".join(
+                get_state_param(self.state, (self._extensions,))
+                if isinstance(self._extensions, str)
+                else self._extensions
+            ),
             classes="d-none",
             ref=self._ref_name,
             # Serialize the content in a way that will work with nova-mvvm and then push it to the server.
@@ -94,10 +109,10 @@ class FileUpload(vuetify.VBtn):
         )
 
         with self:
-            with vuetify.VMenu(activator="parent"):
+            with vuetify.VMenu(v_if=self._show_server_files.expression, activator="parent"):
                 with vuetify.VList():
                     vuetify.VListItem("From Local Machine", click=f"trame.refs.{self._ref_name}.click()")
-                    vuetify.VListItem("From Analysis Cluster", click=self.remote_file_input.open_dialog)
+                    vuetify.VListItem("From Server", click=self.remote_file_input.open_dialog)
 
         @self.server.controller.trigger(f"decode_blob_{self._id}")
         def _decode_blob(contents: bytes) -> None: