+++ tests/gallery/models/file_upload.py
@@ -8,6 +8,6 @@ from pydantic import BaseModel, Field
 class FileUploadState(BaseModel):
     """Model for MVVM demo of FileUpload."""
 
-    extensions: List[str] = Field(default=[".cif", ".nxs"])
+    extensions: List[str] = Field(default=[".cif", ".h5", ".nxs"])
     file: str = Field(default="")
     label: str = Field(default="Upload File")